keycloak and backend JS config moved
