+++ frontend/app/src/views/HomeView.vue
@@ -3,7 +3,7 @@ import TheWelcome from '@/components/TheWelcome.vue'
 import ApiPublicData from '@/components/ApiPublicData.vue'
 import ApiAuthData from '@/components/ApiAuthData.vue'
 import ApiAuthDataWithForm from '@/components/ApiAuthDataWithForm.vue'
-import keycloak from '@/keycloak'
+import keycloak from '@/config/keycloak'
 import ApiUserRoleUser from '@/components/ApiUserRoleUser.vue'
 import ApiUserRoleAdmin from '@/components/ApiUserRoleAdmin.vue'
 import ApiGetUsers from '@/components/ApiGetUsers.vue'
